// src/modules/social/constants/social.enums.ts
export enum FriendRequestStatusEnum {
  PENDING = 'pending',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
  CANCELLED = 'cancelled',
}

export enum GiftStatusEnum {
  SENT = 'sent',
  CLAIMED = 'claimed',
  EXPIRED = 'expired',
}

export enum GiftTypeEnum {
  FGC = 'fgc',
  CHEST = 'chest',
}